moved Moves.php rules to own Rules.php
